@@ -23,33 +23,34 @@
         <project.uri>${project.baseUri}/../</project.uri>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
 
         <twinkleframework.version>0.1.0</twinkleframework.version>
 
-        <spring-cloud.version>Greenwich.SR2</spring-cloud.version>
-        <asm.version>7.1</asm.version>
-        <lombok.version>1.18.8</lombok.version>
-        <guava.version>28.0-jre</guava.version>
-        <commons-text.version>1.7</commons-text.version>
+        <spring-cloud.version>2023.0.2</spring-cloud.version>
+        <asm.version>9.7</asm.version>
+        <lombok.version>1.18.34</lombok.version>
+        <guava.version>33.3.0-jre</guava.version>
+        <commons-text.version>1.12.0</commons-text.version>
         <jregex.version>1.2_01</jregex.version>
         <javamail.version>1.5.0-b01</javamail.version>
-        <commonspool.version>2.6.2</commonspool.version>
-        <commonslang.version>3.8.1</commonslang.version>
-        <commonscollection.version>4.1</commonscollection.version>
-        <commonbeanutils.version>1.9.3</commonbeanutils.version>
-        <jodatime.version>2.10.1</jodatime.version>
-        <logstash.version>6.1</logstash.version>
-        <druid.version>1.1.20</druid.version>
+        <commonspool.version>2.12.0</commonspool.version>
+        <commonslang.version>3.16.0</commonslang.version>
+        <commonscollection.version>4.5.0-M2</commonscollection.version>
+        <commonbeanutils.version>1.9.4</commonbeanutils.version>
+        <jodatime.version>2.12.7</jodatime.version>
+        <logstash.version>8.0</logstash.version>
+        <druid.version>1.2.23</druid.version>
+        <baomidou.version>4.3.1</baomidou.version>
         <p6spy.version>3.8.6</p6spy.version>
-        <mysql-driver.version>8.0.17</mysql-driver.version>
-        <pgsql-driver.version>42.2.8</pgsql-driver.version>
+        <mysql-driver.version>9.0.0</mysql-driver.version>
+        <pgsql-driver.version>42.7.3</pgsql-driver.version>
 
         <hibernatejpa.version>1.0.2.Final</hibernatejpa.version>
         <hibernatevalidator.version>6.0.13.Final</hibernatevalidator.version>
         <validationapi.version>2.0.1.Final</validationapi.version>
-        <springfox.version>2.9.2</springfox.version>
-        <fastjson.version>1.2.60</fastjson.version>
+        <swagger.version>2.6.0</swagger.version>
+        <fastjson.version>2.0.52</fastjson.version>
     </properties>
 
     <dependencyManagement>
