@@ -1,16 +1,13 @@
 package com.twinkle.framework.bootstarter.config;
 
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.servers.Server;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Contact;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.context.request.async.DeferredResult;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 /**
  * Function: TODO ADD FUNCTION. <br/>
@@ -21,28 +18,36 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
  * @see
  * @since JDK 1.8
  */
+@OpenAPIDefinition(
+        servers = {
+                @Server(description = "Dev Server", url = "http://localhost:8080"),
+                @Server(description = "Test Server", url = "https://test.twinkle.com")
+        }
+)
 @Configuration
-@EnableSwagger2
 public class SwaggerConfig {
     @Bean
-    public Docket swaggerSpringMvcPlugin() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .groupName("Standard Service Management")
-                .apiInfo(apiInfo())
-                .genericModelSubstitutes(DeferredResult.class)
-                .useDefaultResponseMessages(false)
-                .forCodeGeneration(true)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.twinkle.framework.bootstarter.controller"))
-                .paths(PathSelectors.any()).build();
+    public OpenAPI openAPI() {
+        return new OpenAPI()
+                // 配置接口文档基本信息
+                .info(this.getApiInfo());
     }
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
+
+    private Info getApiInfo() {
+        return new Info()
+                // 配置文档标题
                 .title("Twinkle Standard Service API.")
-                .description("Twinkle Framework Standard Serivce with Business APIs.")
-                .contact(new Contact("Chenxj", "http://www.google.com", "cxj_hit@126.com"))
-                .version("0.1")
-                .build();
+                // 配置文档描述
+                .description("Twinkle Framework Standard Service with Business APIs.")
+                // 配置作者信息
+                .contact(new Contact().name("Chenxj").url("https://www.github.com").email("cx_hit@126.com"))
+                // 配置License许可证信息
+                .license(new License().name("Apache 2.0").url("https://www.github.com"))
+                // 概述信息
+                .summary("SpringBoot3 Swagger3")
+                .termsOfService("https://www.github.com")
+                // 配置版本号
+                .version("2.0");
     }
 }
 
