@@ -1,19 +1,17 @@
 package com.twinkle.framework.bootstarter.config;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.serializer.SerializerFeature;
-import com.alibaba.fastjson.support.config.FastJsonConfig;
-import com.alibaba.fastjson.support.spring.FastJsonHttpMessageConverter;
+import com.alibaba.fastjson2.JSONWriter;
+import com.alibaba.fastjson2.support.config.FastJsonConfig;
+import com.alibaba.fastjson2.support.spring6.http.converter.FastJsonHttpMessageConverter;
 import com.twinkle.framework.connector.http.converter.TwinkleFastJsonHttpMessageConverter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,55 +28,66 @@ import java.util.List;
 @Configuration
 public class TwinkleWebConfig implements WebMvcConfigurer {
 
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("swagger-ui.html")
-                .addResourceLocations("classpath:/META-INF/resources/");
-        registry.addResourceHandler("/webjars/**")
-                .addResourceLocations("classpath:/META-INF/resources/webjars/");
-    }
+//    @Override
+//    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+//        registry.addResourceHandler("swagger-ui.html")
+//                .addResourceLocations("classpath:/META-INF/resources/");
+//        registry.addResourceHandler("/webjars/**")
+//                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+//    }
 
     @Bean
     public HttpMessageConverters fastJsonHttpMessageConverters() {
-        //1、定义一个convert转换消息的对象
-        FastJsonHttpMessageConverter fastConverter = new TwinkleFastJsonHttpMessageConverter();
-        List<MediaType> fastMediaTypes = new ArrayList<>();
-        fastMediaTypes.add(MediaType.APPLICATION_JSON);
-        fastMediaTypes.add(MediaType.APPLICATION_JSON_UTF8);
-        fastMediaTypes.add(MediaType.APPLICATION_ATOM_XML);
-        fastMediaTypes.add(MediaType.APPLICATION_FORM_URLENCODED);
-        fastMediaTypes.add(MediaType.TEXT_HTML);
-        fastMediaTypes.add(new MediaType("application", "vnd.spring-boot.actuator.v2+json"));
-        fastConverter.setSupportedMediaTypes(fastMediaTypes);
+        // 1. 配置fastjson
+        FastJsonConfig config = new FastJsonConfig();
 
-        FastJsonConfig fastJsonConfig = geFastJsonConfig();
-        //3、在convert中添加配置信息
-        fastConverter.setFastJsonConfig(fastJsonConfig);
-        //4、将convert添加到converters中
-        HttpMessageConverter<?> converter = fastConverter;
+        config.setDateFormat("yyyy-MM-dd HH:mm:ss");
+        config.setCharset(StandardCharsets.UTF_8);
+        config.setWriterFeatures(
+                JSONWriter.Feature.WriteNullListAsEmpty,
+                //json格式化
+                JSONWriter.Feature.PrettyFormat,
+                //输出map中value为null的数据
+                JSONWriter.Feature.WriteMapNullValue,
+                //输出boolean 为 false
+                JSONWriter.Feature.WriteNullBooleanAsFalse,
+                //输出list 为 []
+                JSONWriter.Feature.WriteNullListAsEmpty,
+                //输出number 为 0
+                JSONWriter.Feature.WriteNullNumberAsZero,
+                //输出字符串 为 ""
+                JSONWriter.Feature.WriteNullStringAsEmpty,
+                //对map进行排序
+                JSONWriter.Feature.MapSortField
+        );
+
+        // 2. 添加fastjson转换器
+        FastJsonHttpMessageConverter converter = new TwinkleFastJsonHttpMessageConverter();
+        List<MediaType> supportedMediaTypes = new ArrayList<>();
+
+        // 3. 添加支持的媒体类型
+        supportedMediaTypes.add(MediaType.APPLICATION_JSON);
+        supportedMediaTypes.add(MediaType.APPLICATION_ATOM_XML);
+        supportedMediaTypes.add(MediaType.APPLICATION_FORM_URLENCODED);
+        supportedMediaTypes.add(MediaType.APPLICATION_OCTET_STREAM);
+        supportedMediaTypes.add(MediaType.APPLICATION_PDF);
+        supportedMediaTypes.add(MediaType.APPLICATION_RSS_XML);
+        supportedMediaTypes.add(MediaType.APPLICATION_XHTML_XML);
+        supportedMediaTypes.add(MediaType.APPLICATION_XML);
+        supportedMediaTypes.add(MediaType.IMAGE_GIF);
+        supportedMediaTypes.add(MediaType.IMAGE_JPEG);
+        supportedMediaTypes.add(MediaType.IMAGE_PNG);
+        supportedMediaTypes.add(MediaType.TEXT_EVENT_STREAM);
+        supportedMediaTypes.add(MediaType.TEXT_HTML);
+        supportedMediaTypes.add(MediaType.TEXT_MARKDOWN);
+        supportedMediaTypes.add(MediaType.TEXT_PLAIN);
+        supportedMediaTypes.add(MediaType.TEXT_XML);
+
+        converter.setSupportedMediaTypes(supportedMediaTypes);
+
+        //4 将convert添加到converters
+        converter.setFastJsonConfig(config);
 
         return new HttpMessageConverters(converter);
-    }
-    @Bean
-    public FastJsonConfig geFastJsonConfig() {
-        FastJsonConfig fastJsonConfig = new FastJsonConfig();
-//        SerializeConfig serializeConfig = new SerializeConfig();
-//        NameFilter nameFilter = (object, name, value) -> name;
-//        serializeConfig.addFilter(UiConfiguration.class, nameFilter);
-//        serializeConfig.addFilter(SwaggerResource.class, nameFilter);
-//        serializeConfig.addFilter(MetricsEndpoint.MetricResponse.class, nameFilter);
-//        serializeConfig.addFilter(EnvironmentEndpoint.EnvironmentDescriptor.class, nameFilter);
-//        serializeConfig.propertyNamingStrategy = PropertyNamingStrategy.SnakeCase;
-//        fastJsonConfig.setSerializeConfig(serializeConfig);
-        fastJsonConfig.setDateFormat(JSON.DEFFAULT_DATE_FORMAT);
-//        fastJsonConfig.setSerializerFeatures(SerializerFeature.PrettyFormat);
-
-        fastJsonConfig.setSerializerFeatures(SerializerFeature.WriteMapNullValue,
-                SerializerFeature.WriteNullNumberAsZero,
-                SerializerFeature.WriteNullListAsEmpty,
-                SerializerFeature.WriteNullStringAsEmpty,
-                SerializerFeature.WriteNullBooleanAsFalse,
-                SerializerFeature.WriteDateUseDateFormat);
-        return fastJsonConfig;
     }
 }
