@@ -1,6 +1,6 @@
 package com.twinkle.framework.configure.client.loader;
 
-import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson2.JSON;
 import com.twinkle.framework.configure.client.data.ConfigClientProperties;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
