@@ -1,6 +1,6 @@
 package com.twinkle.framework.configure.component;
 
-import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson2.JSONObject;
 import com.twinkle.framework.api.component.IComponent;
 import com.twinkle.framework.api.component.IComponentFactory;
 import com.twinkle.framework.api.config.Configurable;
