@@ -1,6 +1,6 @@
 package com.twinkle.framework.connector.http.endpoint;
 
-import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson2.JSONObject;
 import com.twinkle.framework.api.component.AbstractComponent;
 import com.twinkle.framework.api.constant.ExceptionCode;
 import com.twinkle.framework.api.exception.ConfigurationException;
@@ -139,7 +139,7 @@ public abstract class AbstractHttpEndpoint extends AbstractComponent implements 
      */
     private Set<String> getAnnotaions() {
         Set<String> tempAnnotations = new HashSet<>();
-        tempAnnotations.add("@io.swagger.annotations.ApiOperation(value = \"" + this.description + "\")");
+        tempAnnotations.add("@io.swagger.v3.oas.annotations.Operation(summary = \"" + this.description + "\")");
         switch (this.requestType) {
             case GET:
                 tempAnnotations.add("@org.springframework.web.bind.annotation.RequestMapping(value = \"" + this.url + "\", method = org.springframework.web.bind.annotation.RequestMethod.GET)");
