@@ -1,6 +1,6 @@
 package com.twinkle.framework.connector.http.endpoint;
 
-import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson2.JSONObject;
 import com.twinkle.framework.api.exception.ConfigurationException;
 import com.twinkle.framework.asm.descriptor.TypeDescriptor;
 import com.twinkle.framework.asm.descriptor.TypeDescriptorImpl;
