@@ -1,6 +1,6 @@
 package com.twinkle.framework.datacenter.statement;
 
-import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson2.JSONObject;
 import com.twinkle.framework.api.constant.ExceptionCode;
 import com.twinkle.framework.api.exception.ConfigurationException;
 import lombok.extern.slf4j.Slf4j;
