@@ -1,7 +1,7 @@
 package com.twinkle.framework.ruleengine;
 
-import com.alibaba.fastjson.JSONArray;
-import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson2.JSONArray;
+import com.alibaba.fastjson2.JSONObject;
 import com.twinkle.framework.api.component.AbstractComponent;
 import com.twinkle.framework.api.component.rule.IRule;
 import com.twinkle.framework.api.component.rule.IRuleChain;
